1.Added validation for command line parameters 2.Refactored code

+++ src/main/java/com/ooyala/playback/page/EndScreenValidator.java
@@ -46,9 +46,9 @@ public class EndScreenValidator extends PlayBackPage implements
 			return false;
 		}
 
-		double currenttime = Double.parseDouble(((JavascriptExecutor) driver)
+		int currenttime = Integer.parseInt(((JavascriptExecutor) driver)
 				.executeScript("return pp.getPlayheadTime().toFixed();").toString());
-		double totaltime = Double.parseDouble(((JavascriptExecutor) driver)
+		int totaltime = Integer.parseInt(((JavascriptExecutor) driver)
 				.executeScript("return pp.getDuration().toFixed();").toString());
 		
 		if(!(getBrowser().equalsIgnoreCase("internet explorer") || getBrowser().equalsIgnoreCase("MicrosoftEdge") || (getBrowser().equalsIgnoreCase("firefox") && getPlatform().equalsIgnoreCase("windows")))){
